requirements.txt
langchain
langchain-core
langchain-community
langchain-text-splitters
langchain-google-genai
sentence-transformers
faiss-cpu
rank_bm25
pypdf
streamlit
python-dotenv
google-generativeai
scikit-learn
bert-score
evaluate
numpy
